Extract ShopSession.context module
